+++ example/servo.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var Gpio = require('../'),
+var Gpio = require('../').Gpio,
   motor = new Gpio(10, {mode: Gpio.OUTPUT}),
   pulseWidth = 1000,
   increment = 100;