+++ test/pull-up-down.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var assert = require('assert'),
-  Gpio = require('../'),
+  Gpio = require('../').Gpio,
   input = new Gpio(22, {mode: Gpio.INPUT, pullUpDown: Gpio.PUD_UP});
 
 assert.strictEqual(input.digitalRead(), 1, 'expected gpio22 to be 1');