+++ test/pwm.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var assert = require('assert'),
-  Gpio = require('../'),
+  Gpio = require('../').Gpio,
   iv,
   led = new Gpio(17, {mode: Gpio.OUTPUT}),
   dutyCycle = 0;