+++ History.md
@@ -1,3 +1,8 @@
+Unpublished
+===========
+
+  * export Gpio constructor as exports.Gpio
+
 0.0.8 / Nov 08 2015
 ===================